initial work to have topological cannonfile
still need to make a couple changes for contxet handling and stuff, but
overall shaping really beautifully.

also, users of cannon need to be made aware that they need to mark a
steps as linearlly dependant if they both depend on the same memory
storage slots. Othrewise, there is some potential for corruption of
storage