add abiOf to `factory` on invoke
if you have a contract which is generated and is not available in the
artifacts, it may be necessary to pull the ABI from a previously deployed
contract instead.

this could be a problem for etherscan verification because in this case cannon
doesn't technically have the source in order to verify the contract.

not sure about hte naming. Thoughts?